--- a/dependency_range_tests/scripts/langchain/node/update_resolutions_lowest.js
+++ b/dependency_range_tests/scripts/langchain/node/update_resolutions_lowest.js
@@ -3,9 +3,14 @@ const semver = require("semver");
 
 const communityPackageJsonPath = "package.json";
 
-const currentPackageJson = JSON.parse(fs.readFileSync(communityPackageJsonPath));
+const currentPackageJson = JSON.parse(
+  fs.readFileSync(communityPackageJsonPath)
+);
 
-if (currentPackageJson.peerDependencies["@langchain/core"] && !currentPackageJson.peerDependencies["@langchain/core"].includes("rc")) {
+if (
+  currentPackageJson.peerDependencies?.["@langchain/core"] &&
+  !currentPackageJson.peerDependencies["@langchain/core"].includes("rc")
+) {
   const minVersion = semver.minVersion(
     currentPackageJson.peerDependencies["@langchain/core"]
   ).version;
@@ -15,11 +20,14 @@ if (currentPackageJson.peerDependencies["@langchain/core"] && !currentPackageJso
   };
 }
 
-if (currentPackageJson.devDependencies["@langchain/core"]) {
+if (currentPackageJson.devDependencies?.["@langchain/core"]) {
   delete currentPackageJson.devDependencies["@langchain/core"];
 }
 
-if (currentPackageJson.dependencies["@langchain/openai"] && !currentPackageJson.dependencies["@langchain/openai"].includes("rc")) {
+if (
+  currentPackageJson.dependencies?.["@langchain/openai"] &&
+  !currentPackageJson.dependencies["@langchain/openai"].includes("rc")
+) {
   const minVersion = semver.minVersion(
     currentPackageJson.dependencies["@langchain/openai"]
   ).version;
@@ -29,7 +37,10 @@ if (currentPackageJson.dependencies["@langchain/openai"] && !currentPackageJson.
   };
 }
 
-if (currentPackageJson.dependencies["@langchain/textsplitters"] && !currentPackageJson.dependencies["@langchain/textsplitters"].includes("rc")) {
+if (
+  currentPackageJson.dependencies?.["@langchain/textsplitters"] &&
+  !currentPackageJson.dependencies["@langchain/textsplitters"].includes("rc")
+) {
   const minVersion = semver.minVersion(
     currentPackageJson.dependencies["@langchain/textsplitters"]
   ).version;
@@ -42,7 +53,10 @@ if (currentPackageJson.dependencies["@langchain/textsplitters"] && !currentPacka
 // Stupid hack
 currentPackageJson.resolutions = {
   ...currentPackageJson.resolutions,
-  "jackspeak": "2.1.1"
+  jackspeak: "2.1.1",
 };
 
-fs.writeFileSync(communityPackageJsonPath, JSON.stringify(currentPackageJson, null, 2));
+fs.writeFileSync(
+  communityPackageJsonPath,
+  JSON.stringify(currentPackageJson, null, 2)
+);