--- a/dependency_range_tests/scripts/with_standard_tests/community/test-with-lowest-deps.sh
+++ b/dependency_range_tests/scripts/with_standard_tests/community/test-with-lowest-deps.sh
@@ -35,7 +35,7 @@ yarn
 cd "$monorepo_community_dir"
 
 # Read the @langchain/core version from peerDependencies
-core_version=$(node -p "require('./package.json').peerDependencies['@langchain/core']")
+core_version=$(node -p "require('./package.json').peerDependencies?.['@langchain/core']")
 
 # Install @langchain/core at the specified version
 yarn add @langchain/core@$core_version