--- a/environment_tests/test-exports-vite/vite.config.js
+++ b/environment_tests/test-exports-vite/vite.config.js
@@ -1,0 +1,7 @@
+export default {
+  build: {
+    rollupOptions: {
+      external: [/^node:/, "typeorm", "reflect-metadata"],
+    },
+  },
+};