--- a/langchain-core/src/utils/types/zod.ts
+++ b/langchain-core/src/utils/types/zod.ts
@@ -11,7 +11,6 @@ import {
   $ZodUnknown,
   $ZodNever,
   $ZodOptional,
-  _array,
 } from "zod/v4/core";
 
 export type ZodStringV3 = z3.ZodString;