## Overview

This PR introduces a completely rewritten build system for the LangChain TypeScript monorepo, replacing the previous Rollup-based approach with a modern, plugin-based architecture built on [tsdown](https://github.com/privatenumber/tsdown) and Rolldown.

## Motivation

My main motivation working on this has been the need for faster build times especially when working on multiple packages in parallel. I was missing a `watch` command that allowed me to just keep a terminal window open in the background and forget about it. I've implemented a similar build architecture in WebdriverIO (using Esbuild) that turned out to be very efficient, however today there have been a lot of advancements around the Rolldown project, making the original build system I had in mind even more faster and efficient.

### Architecture & Technology

| Aspect | Old System (Rollup) | New System (tsdown) |
|--------|------------------|---------------------|
| **Bundler** | Rollup | tsdown + Rolldown |
| **Architecture** | Monolithic script | Plugin-based architecture |
| **Configuration** | `langchain.config.js` per package | Centralized `infra/build/` |
| **Code Generation** | Manual script execution | Automated plugins |
| **Type Checking** | Basic TypeScript | Advanced validation with ATTW, publint |
| **Compiler** | `typescript@v5` | `@typescript/native-preview@7.0.0-dev.20250616.1` |

### Key Improvements

#### 🚀 **Performance & Developer Experience**
- **Faster builds**: tsdown with Rolldown provides significantly better performance than TSup
- **Watch mode**: Real-time compilation with intelligent file watching
- **Parallel processing**: Builds multiple packages simultaneously
- **Rich CLI**: Comprehensive command-line interface with flexible filtering options

#### 🔧 **Build Quality & Validation**
```diff
Old System:
- Basic TypeScript compilation
- Manual tree-shaking checks with complex rollup setup

New System:
+ TypeScript Go compilation
+ Integrated type checking with arethetypeswrong
+ Built-in publint validation
+ Automated unused dependency detection
```

#### 📦 **Package Management**
```diff
Old System:
- Manual entrypoint generation
- Complex file moving and renaming logic
- Hardcoded path transformations
- Unnecessary support for Node Classic imports

New System:
+ Automatic workspace discovery
+ Intelligent package filtering
+ Simplified export structure with `input` field - removes the need for custom `langchain.config.js`
+ Plugin-based code generation
```

### Plugin System

The current build setup has a custom implementation to generate, `lc-secrets.ts`, `import-map.ts` and `import-constants.ts`. I've moved this into a Rolldown plugin, which provides better modularity and abstraction from the build system.

### Export Structure Modernization

**Old System**: Complex manual file generation
```json
{
  "exports": {
    ".": {
      "types": "./index.d.ts",
      "import": "./index.js", 
      "require": "./index.cjs"
    }
  }
}
```

**New System**: Input-driven exports with automatic builds
```json
{
  "exports": {
    ".": {
      "input": "./src/index.ts",  // ← New: Source file reference
      "require": {
        "types": "./dist/index.d.cts",
        "default": "./dist/index.cjs"
      },
      "import": {
        "types": "./dist/index.d.ts",
        "default": "./dist/index.js"
      }
    }
  }
}
```

## 🎯 Build System Features

### Command Line Interface
```bash
# Build all packages
yarn build

# Watch mode for development
yarn build --watch

# Target specific packages
yarn build @langchain/core langchain

# Exclude packages
yarn build --exclude @langchain/community

# Skip various build steps
yarn build --no-emit --skip-unused --skip-sourcemap
```

## 🔄 Migration Impact

### For Developers
- **Simplified workflow**: Single `yarn build:new` command replaces complex build scripts
- **Faster iteration**: Watch mode provides instant feedback and rebuilding of files
- **Consistent output**: Standardized build artifacts across all packages

### For CI/CD
- **Performance gains**: Faster build times with modern tooling
- **Quality gates**: Built-in validation prevents problematic releases
- **Selective building**: Only build what's changed

### For LangChain Users
- no changes expected

## 🛣️ Migration Status

- ✅ **Infrastructure**: Complete new build system implemented
- ✅ **Core packages**: langchain-core and langchain updated
- ✅ **All libs**: All `libs/langchain-*` packages migrated
- ✅ **Code generation**: All import maps, constants, and types regenerated
- ⏳ **Testing**: Comprehensive testing in progress
- ⏳ **Documentation**: Build system documentation being finalized

## 🏃 Performance Analysis

I've measured the build times of the old vs new build system.

| Scenario | Old System (Rollup) | New System (tsdown) |
|--------|------------------|---------------------|
| **Complete Build (CJS + ESM + d.ts + SourceMap)** | ~144s (via `yarn build`) | ~23s (via `yarn build`) |
| **Single Package build (CJS + ESM + d.ts + SourceMap)** | 13.716s (via `npx turbo build --filter=@langchain/core`) | ~2.5s (via `yarn build @langchain/core`) |
| **NoEmit Build (CJS + ESM)** | n/a | ~4.1s (via `yarn build --noEmit`) |
| **Rebuild after file change** | n/a | ~1.5s (via `yarn build --watch`) |

## 📝 Notes

This is a **work in progress** and represents a major infrastructure improvement for the LangChain project. The new system provides better performance, reliability, and developer experience while maintaining full compatibility with existing packages.