@christian-bromann 2025-08-01
Apologize @hntrl for this mess, this got bigger than expected. I tried to remove all changes in files that I didn't had to touch. I went through every change and noted additional changes coming with this PR in addition to the PR description. Happy to walk through them in person.

## 📦 CI
- updated invalid syntax in workflow files:
    ```patch
    - paths_ignore:
    + paths-ignore:
    ```
- updated `actions/setup-node` action versions from `v3` to `v4`
- updated Node.js version in workflows from `18.x` to `20.x`
- made `actions/setup-node` action get its Node.js version from `.nvmrc` file
- re-enabled Bun environment tests in `test-exports.yml` workflow

## 🧹 Cleanup
- contains prettier changes that have been ignored in the past

## 📝 Documentation
- skip build command to speed up CI as we plan to move docs over to a new repo (see `docs/core_docs/package.json`)

## 🧪 Tests
- updated environment tests to proper setup dependencies for `langchain` after peer dependency update
- ⚠️ I had to create a `next.config.js` file in the `langchain` package to make the tests pass.
  This is due to the fact that peer dependencies like `typeorm` which is a Node.js only package were
  resolved by Next.js to be bundled into a browser package. This will be obsolete once we update
  the `langchain` package and remove the SQL primitives.

## Examples
- moved creation of tool schemas out of class initialization as `z.infer<typeof this.schema>` was
  not working anymore and failed during compilation.

## 📦 Packages

- the export pattern of package.json files has changed from
  ```json
  "./caches": {
    "types": {
      "import": "./caches.d.ts",
      "require": "./caches.d.cts",
      "default": "./caches.d.ts"
    },
    "import": "./caches.js",
    "require": "./caches.cjs"
  }
  ```
  to:
  ```json
  "./caches": {
    "require": {
      "types": "./dist/caches/base.d.cts",
      "default": "./dist/caches/base.cjs"
    },
    "import": {
      "types": "./dist/caches/base.d.ts",
      "default": "./dist/caches/base.js"
    },
    "input": "./src/caches/base.ts"
  },
  ```
- removed all entries except `dist` out of the "files" field
- `import_map.ts` files were updated and are now actually matching with what
  is being defined in the `exports` field
- in several files I added return types to functions that were missing and were
  causing compilation errors (TypeScript can't refer them automatically when
  they come from an external package)
- removed several barrel files e.g. `langchain/src/cache/base.ts` that would only
  contain single exports like `export * from "@langchain/core/caches";`
- replaced a lot of barrel imports with direct ones, e.g.
  ```patch
  - import { BaseDocumentLoader } from "../base.js";
  + import { BaseDocumentLoader } from "@langchain/core/document_loaders/base";
  ```

### 🧳 `langchain`

- in `langchain/src/chat_models/universal.ts` I replaced direct dynamic imports
  like `import("@langchain/anthropic")` with `import(config.package)` and created
  a custom mapping - this allowed me to remove the peer dependency to various
  internal packages
- transformed ambient definition of `expression-parser` to a proper type definition
  to simplify imports
- removed usage of `axios` in `langchain/src/tools/webbrowser.ts` in favor of native
  fetch to remove dependency on it