@christian-bromann 2025-08-01
> would be neat if we just implicitly included `load/import_map.ts` as a build artifact rather than checking it into source control

Not sure if that is feasible as we directly import them in e.g. `langchain/src/load/index.ts`.